support probabilistic control of trip distance. reorg commands into files
new file:   cmds/__init__.py
new file:   cmds/add_location.py
new file:   cmds/get_costs.py
modified:   main.py
