+++ cmds/__init__.py
@@ -1,0 +1,2 @@
+from .add_location import add_location
+from .get_costs import get_costs
